requirements.txt
asgiref==3.2.10
boto3==1.16.44
botocore==1.19.44
cffi==1.14.3
cryptography==3.1.1
dj-database-url==0.5.0
Django==3.1
django-disqus==0.5
django-rest-knox==4.1.0
django-storages==1.11.1
djangorestframework==3.12.1
gunicorn==20.0.4
jmespath==0.10.0
mock==4.0.3
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
whitenoise==5.2.0
